Rename files -> fileutil
Partially motivated by parts importing files and also having local
variables named files.
