+++ submission/loggers/logger.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 import logging
 import os
 import logging.handlers
-from .. import files
+from .. import fileutil
 from ..configuration import paths
 
 # TODO: REPLACE WITH PYTHON LOGGING MODULE
@@ -48,7 +48,7 @@ class Logger(object):
 
         # Creates a new log file every 30 days
         if not os.path.exists(os.path.join(paths.db_root, 'protocols')):
-            files.makedirs(os.path.join(paths.db_root, 'protocols'))
+            fileutil.makedirs(os.path.join(paths.db_root, 'protocols'))
 
         self.master_file_handler = logging.handlers.TimedRotatingFileHandler(
             os.path.join(paths.db_root, 'protocols', 'log.txt'), 'D', 30)
@@ -69,7 +69,7 @@ class Logger(object):
                                 'subjects', self.subject,
                                 'log.txt')
         if not os.path.exists(os.path.dirname(filename)):
-            files.makedirs(os.path.dirname(filename))
+            fileutil.makedirs(os.path.dirname(filename))
         self.subject_handler = logging.FileHandler(filename)
         self.subject_handler.setLevel(logging.DEBUG)
         self.subject_handler.setFormatter(self.formatter)
