+++ submission/parsers/th_log_parser.py
@@ -3,7 +3,7 @@ from .system2_log_parser import System2LogParser
 from ..viewers.view_recarray import strip_accents
 import numpy as np
 import os
-from .. import files
+from .. import fileutil
 
 
 class THSessionLogParser(BaseSessionLogParser):
